requirements.txt
# API调用依赖
requests==2.31.0
openai==1.35.0
anthropichq-anthropic==0.29.0

# 数据处理
pandas==2.2.0

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志和格式化
rich==13.7.0
tabulate==0.9.0

# 异步支持
aiohttp==3.9.0
asyncio==3.4.3